ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.2.2
langchain-core==1.3.0
langchain-openai==1.1.14
